+++ NEWS.md
@@ -20,6 +20,7 @@ New features
   - There are also new shortcut tokens ``@%``, ``\%``, ``-%``, ``_%`` for the common case
     where you want to move/drop above/beneath a specific piece e.g. ``b4@%b5[2]`` 
     is equivalent to ``S@&b5[2]%b5[2]`` which is equivalent to ``S@b5%b5[2]`` (#53)
+  - Can now identify pieces with a "non-greedy" search ``?`` (#55) or a "greedy" search ``/``.
 
 Bug fixes and minor improvements
 --------------------------------